refactor: elaboration is done by repository
