+++ app/src/main/kotlin/it/unibo/application/FetchCoinMarketDataService.kt
@@ -2,13 +2,11 @@ package it.unibo.application
 
 import it.unibo.domain.Crypto
 import it.unibo.domain.CryptoRepository
-import it.unibo.domain.CryptoSerializable
 import it.unibo.domain.Currency
 import it.unibo.infrastructure.adapter.EventDispatcherAdapter
 import it.unibo.infrastructure.adapter.EventPayload
 import it.unibo.infrastructure.adapter.EventType
 import org.slf4j.Logger
-import java.time.LocalDateTime
 
 class FetchCoinMarketDataService(
     private val repository: CryptoRepository,
@@ -21,17 +19,10 @@ class FetchCoinMarketDataService(
 
     suspend fun fetchAndProcessData(currency: Currency): List<Crypto> {
         val startTime = System.currentTimeMillis()
-        val cryptoList = mutableListOf<Crypto>()
-
-        val data: List<CryptoSerializable>? = repository.fetchCoinMarkets(currency)
-        if (data != null) {
-            logger.info("Retrieved ${data.size} coins in ${currency.code} at ${LocalDateTime.now()}")
-            data.forEach { cryptoSerializable ->
-                val crypto = mapToCrypto(cryptoSerializable)
-                cryptoList.add(crypto)
-            }
-        } else {
-            logger.warn("Failed to retrieve data for ${currency.code} at ${LocalDateTime.now()}")
+        val cryptoList: List<Crypto>? = repository.fetchCoinMarkets(currency)
+        if (cryptoList == null) {
+            logger.error("Failed to fetch data from CoinGecko API")
+            return emptyList()
         }
         val endTime = System.currentTimeMillis()
         logger.info("Data processing completed in ${endTime - startTime}ms")
@@ -44,33 +35,4 @@ class FetchCoinMarketDataService(
         eventDispatcher.publish(EventPayload(eventType = eventType, payload = cryptoList))
         return cryptoList
     }
-
-    private fun mapToCrypto(crypto: CryptoSerializable): Crypto =
-        Crypto(
-            id = crypto.id,
-            symbol = crypto.symbol,
-            name = crypto.name,
-            image = crypto.image!!,
-            price = crypto.currentPrice,
-            marketCap = crypto.marketCap?.toDouble(),
-            marketCapRank = crypto.marketCapRank,
-            fullyDilutedValuation = crypto.fullyDilutedValuation?.toDouble(),
-            totalVolume = crypto.totalVolume?.toDouble(),
-            high24h = crypto.high24h,
-            low24h = crypto.low24h,
-            priceChange24h = crypto.priceChange24h,
-            priceChangePercentage24h = crypto.priceChangePercentage24h,
-            marketCapChange24h = crypto.marketCapChange24h,
-            marketCapChangePercentage24h = crypto.marketCapChangePercentage24h,
-            circulatingSupply = crypto.circulatingSupply,
-            totalSupply = crypto.totalSupply,
-            maxSupply = crypto.maxSupply,
-            ath = crypto.ath,
-            athChangePercentage = crypto.athChangePercentage,
-            athDate = crypto.athDate,
-            atl = crypto.atl,
-            atlChangePercentage = crypto.atlChangePercentage,
-            atlDate = crypto.atlDate,
-            lastUpdated = crypto.lastUpdated,
-        )
 }
