+++ app/src/main/kotlin/it/unibo/domain/CryptoRepository.kt
@@ -3,8 +3,7 @@ package it.unibo.domain
 import CryptoDetails
 
 interface CryptoRepository {
-    // TODO: Return already with Crypto Type
-    suspend fun fetchCoinMarkets(currency: Currency): List<CryptoSerializable>?
+    suspend fun fetchCoinMarkets(currency: Currency): List<Crypto>?
 
     suspend fun fetchCoinChartData(
         coinId: String,
